third_party/blink/web_tests/wpt_internal/css/css-contain/container-queries/container-type-scroll-state-computed.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Containment Test: Computed values of container-type with scroll-state</title>
<link rel="help" href="https://drafts.csswg.org/css-contain-4/#container-type">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<script src="/css/css-contain/container-queries/support/cq-testcommon.js"></script>
<div id="target"></div>
<script>
  setup(() => assert_implements_container_queries());

  test_computed_value('container-type', 'scroll-state');
  test_computed_value('container-type', 'scroll-state size', 'size scroll-state');
  test_computed_value('container-type', 'inline-size scroll-state');
</script>
